+++ docs/experiments/raven-target/CoreDeclarations.cs
@@ -54,7 +54,7 @@ static class CoreDeclarations
                     info.Methods.Add(factory);
                 }
             }
-            if (collectionProbe) { IntrospectionHierarchy.Project(module); TaskBindings.Project(module); WorkerBindings.Project(module); }
+            if (collectionProbe) { IntrospectionHierarchy.Project(module); TaskBindings.Project(module); WorkerBindings.Project(module); ReaderBindings.Project(module); }
             StorageHierarchy.Project(module);
             NamespaceFunctions.ProjectMath(module);
             NamespaceFunctions.ProjectFault(module);