+++ scripts/build-api-docs.py
@@ -24,6 +24,8 @@ def inputs():
     paths += list((ROOT / 'runtime/raven/src/System/Tasks').glob('*.rvn'))
     paths += list((ROOT / 'runtime/raven/src/System/Concurrency').glob('*.rvn'))
     paths += list((ROOT / 'runtime/raven/src/System/IO').glob('*.rvn'))
+    paths += list((ROOT / 'runtime/raven/src/System/Console').glob('*.rvn'))
+    paths += [ROOT / 'runtime/raven/src/System/ConsoleReadError.rvn']
     paths += list((ROOT / 'runtime/raven/src/System/Storage').rglob('*.rvn'))
     paths += [DOCS / name for name in ('NeoCLR.CoreProbe.xml', 'filter.yml', 'docfx.json')]
     paths += [ROOT / '.config/dotnet-tools.json']